requirements.txt
pytest
uvicorn[standard]
face_recognition
fastapi
coverage
psycopg2-binary